to long double

+++ is_prime_prog/is_prime.h
@@ -1,3 +1,3 @@
 #pragma once
 
-bool is_prime(int number);
+bool is_prime(long double number);
